Information Gain

Information Gain is a feature selection technique based on information theory. It measures the information obtained for the target variable by knowing the presence or absence of a feature. It wraps the FSelector library.

Usage

fs_ig(attribute)

Arguments

attribute

The target variable.

Value

A fs_ig object.

Examples

data(iris)
myfeature <- daltoolbox::fit(fs_ig("Species"), iris)
data <- daltoolbox::transform(myfeature, iris)
head(data)
#>   Petal.Width Petal.Length Species
#> 1         0.2          1.4  setosa
#> 2         0.2          1.4  setosa
#> 3         0.2          1.3  setosa
#> 4         0.2          1.5  setosa
#> 5         0.2          1.4  setosa
#> 6         0.4          1.7  setosa