Gaussian Naive Bayes Classifier

Implements classification using the Gaussian Naive Bayes algorithm. This function wraps the GaussianNB from Python's scikit-learn library.

Usage

skcla_nb(attribute, slevels, var_smoothing = 1e-09, priors = NULL)

Arguments

attribute

Target attribute name for model building

slevels

List of possible values for classification target

var_smoothing

Portion of the largest variance of all features that is added to variances

priors

Prior probabilities of the classes. If specified must be a list of length n_classes

Value

A Naive Bayes classifier object

skcla_nb object

Details

Naive Bayes Classifier

Examples

#See an example of using `skcla_nb` at this
#https://github.com/cefet-rj-dal/daltoolboxdp/blob/main/examples/skcla_nb.md